feat: improve tree metrics sidebar styles
